Border size to float, add min and max size, other small changes

+++ osu.Game/Skinning/SkinConfiguration.cs
@@ -25,7 +25,7 @@ namespace osu.Game.Skinning
 
         public int HitCircleOverlap { get; set; }
 
-        public int? SliderBorderSize { get; set; }
+        public float? SliderBorderSize { get; set; }
 
         public bool? CursorExpand { get; set; } = true;
     }
